move ServiceParams to JibriService
